requirements.txt
browser-cookie3
requests
pypiwin32
pywget